+++ src/main/scala/com/actelion/research/arcite/core/transforms/cluster/workers/fortest/WorkExecUpperCase.scala
@@ -28,7 +28,7 @@ class WorkExecUpperCase extends Actor with ActorLogging with ArciteJSONProtocol 
       val end = java.util.concurrent.ThreadLocalRandom.current().nextInt(10, 100)
       (1 to end).foreach { e ⇒
         Thread.sleep(5000)
-        sender() ! WorkerProgress(e.toDouble * 100.0 / end.toDouble)
+        sender() ! WorkerProgress(e * 100/ end)
       }
       log.info("waited enough time, doing the work now...")
